+++ src/main/java/com/dku/council/domain/with_dankook/repository/with_dankook/BearEatsRepository.java
@@ -1,4 +1,4 @@
-package com.dku.council.domain.with_dankook.repository;
+package com.dku.council.domain.with_dankook.repository.with_dankook;
 
 import com.dku.council.domain.with_dankook.model.entity.type.BearEats;
 import org.springframework.data.domain.Page;
